move run disclaimed logic into allowedcmd package
